+++ doc/changelog.rst
@@ -9,10 +9,20 @@ Unrealeased
   When releasing next version, the "Unreleased" header will be replaced
   with appropriate version header and this help text will be removed.
 
+Admin
+~~~~~
+
+- Add `HexColorWidget` to pick hexadecimal colors on input fields
+
 General
 ~~~~~~~
 
 - Add `front_model_url_resolver` and `admin_model_url_resolver` provides key to resolve models URLs
+
+Carousel
+~~~~~~~~
+
+- Add option to customize arrows and slide dot colors
 
 Shuup 1.6.5
 -----------